+++ tests/database/form_queries/get_form.test.js
@@ -1,6 +1,6 @@
 const test = require('tape');
-const dbReset = require('../../src/model/database/db_seed');
-const { getForm } = require('../../src/model/form_queries');
+const dbReset = require('../../../src/model/database/db_seed');
+const { getForm } = require('../../../src/model/form_queries');
 
 test('Get form from database', (t) => {
   // TODO hardcoding id may be brittle
