+++ tests/database/user_queries/add_user.test.js
@@ -1,7 +1,7 @@
 const test = require('tape');
-const dbConnection = require('../../src/model/database/db_connection');
-const dbReset = require('../../src/model/database/db_seed');
-const { addUser } = require('../../src/model/user_queries');
+const dbConnection = require('../../../src/model/database/db_connection');
+const dbReset = require('../../../src/model/database/db_seed');
+const { addUser } = require('../../../src/model/user_queries');
 
 test('Insert user into database', (t) => {
   dbReset()
